backend/requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.3
mysql-connector-python==8.4.0 